requirements.txt
# COCA Timer - PyQt6 Floating Overlay Timer
# Clean, minimal dependencies for stable overlay

# UI Framework
PyQt6>=6.5.0

# OCR Dependencies
pytesseract>=0.3.10
Pillow>=10.0.0

# Screenshot capture
mss>=9.0.1

# Image enhancement
opencv-python>=4.8.0

# Keyboard input detection
pynput>=1.7.6

# Image processing
numpy>=1.24.0
